requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==4.1.5
gunicorn==20.1.0
Pillow==10.2.0
pip==23.3.2
postgres
djangorestframework
markdown
django-filter